【服务器】 - 增加客户端查询已有角色方法

+++ Server/YouYouServer/YouYouServer.Commmon/DBData/RoleEntity.cs
@@ -24,12 +24,12 @@ namespace YouYouServer.Common
         /// <summary>
         /// 职业编号
         /// </summary>
-        public int JobId;
+        public byte JobId;
 
         /// <summary>
         /// 性别
         /// </summary>
-        public int Sex;
+        public byte Sex;
 
         /// <summary>
         /// 昵称
